feat: implement Contrato model and service with decimal-based quantity handling and database migration

+++ src/models/Contrato.js
@@ -60,7 +60,7 @@ class Contrato extends Model {
           },
         },
         quantidade: {
-          type: DataTypes.INTEGER,
+          type: DataTypes.DECIMAL(10, 1),
           validate: {
             min: {
               args: [0],
